Revert 109855 - Fix quota value overflow.
BUG=104111
TEST='QuotaManagerTest.*'


Review URL: http://codereview.chromium.org/8560007

TBR=tzik@chromium.org
Review URL: http://codereview.chromium.org/8528021

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@109856 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: ae4bb64d5dbfc0470a83c32afd06e864b22bd384

+++ webkit/quota/quota_manager.cc
@@ -755,7 +755,7 @@ class QuotaManager::UpdatePersistentHostQuotaTask
   UpdatePersistentHostQuotaTask(
       QuotaManager* manager,
       const std::string& host,
-      int64 new_quota,
+      int new_quota,
       const HostQuotaCallback& callback)
       : DatabaseTaskBase(manager),
         host_(host),
